requirements2.txt
arxiv==2.0.0
memgpt==0.1.0
openai==1.3.7
pyautogen==0.2.0
